syn: Factor out target index for reuse
